pyproject.toml
[project]
name = "real-estate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=6.0.1",
    "django-environ>=0.12.0",
    "django-filter>=25.2",
    "django-jazzmin>=3.0.1",
    "django-leaflet>=0.33.0",
    "django-prometheus>=2.4.0",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "psycopg[binary]>=3.3.2",
    "python-dotenv>=1.2.1",
]
